requirements.txt
poetry==1.7.1
fastapi==0.110.0
python-multipart==0.0.9
uvicorn==0.27.1
jinja2==3.1.3
python-dotenv==1.0.0
git+https://github.com/landing-ai/agentic-doc.git
supabase==2.15.0
python-jose[cryptography]==3.4.0
requests==2.32.3 
openai==1.75.0